==> notebooks/3_feature_engineering.ipynb <==
# Imports

import pandas as pd
import pandas_profiling
import numpy as np
import seaborn as sns
import matplotlib.pyplot as plt
from pathlib import Path

pd.options.display.max_rows = 80

# Data 

data_path = Path('../data/project')
X = pd.read_csv(data_path/'X_train_preprocessed.csv', index_col = 'Id')
y = pd.read_csv(data_path/'y_train_preprocessed.csv', index_col = 'Id')
X_test = pd.read_csv(data_path/'X_test_preprocessed.csv', index_col = 'Id')

X_full = pd.read_csv(data_path/'train.csv', index_col = 'Id')

# Feature engineering
* Total frontage 

# Finding out which features are not standard at houses

mask = [False if i == 0 else True for i in X_full.isna().sum()]
has_na_cols = X_full.columns[mask]
has_na_cols

def create_features(X):

    X = X.copy()

    # Total square feet
    X['TotalSF'] = X['TotalBsmtSF'] + X['1stFlrSF'] + X['2ndFlrSF']

    # Boolean for if the property has a various features
    feats = ['LotFrontage', 'Alley', 'MasVnrType', 'BsmtQual', 'Electrical', 'FireplaceQu', 'GarageType', 'PoolQC', 'Fence',
    'MiscFeature']
    for feature in feats:
        X['has_'+feature] = X_full[feature].apply(lambda x: 1 if x == x else 0)

    return X

X_engineered = create_features(X)
X_test_engineered = create_features(X_test)

# Testing to see if the binary 'has_feature' engineering worked

feats = ['LotFrontage', 'Alley', 'MasVnrType', 'BsmtQual', 'Electrical', 'FireplaceQu', 'GarageType', 'PoolQC', 'Fence',
    'MiscFeature']
for i in feats:
    print(f'{i} percentage that isn\'t NA:', X_engineered['has_'+i].sum() / X_engineered.shape[0]*100)

X_train.info("")

# Baseline linear model

from sklearn.ensemble import RandomForestRegressor
from sklearn.model_selection import train_test_split


X_train, X_val, y_train, y_val = train_test_split(X_engineered, y, test_size = 0.2)

model_rf = RandomForestRegressor(n_jobs=-1)
model_rf.fit(X_train, y_train)

# Getting baseline RMSE

from sklearn.metrics import mean_squared_error as mse

y_pred = model_rf.predict(X_val)

# Reversing the log target
y_pred = np.exp(y_pred)

y_pred_train = model_rf.predict(X_train)
y_pred_train = np.exp(y_pred_train)

y_val = np.exp(y_val)
rmse_baseline = mse(y_val, y_pred, squared=False)
rmse_train = mse(np.exp(y_train), y_pred_train, squared=False)

print(f'The training RMSE of my RF baseline is: {rmse_train}')
print(f'The validation RMSE of my RF baseline is: {rmse_baseline}')

X_val

# Checking out the feature premutation importances
import eli5
from eli5.sklearn import PermutationImportance

permuter = PermutationImportance(
    model_rf,
    n_iter = 5
)
permuter.fit(X_val, y_val)

# Viewing the top permutation importance features
feature_names = X_train.columns.to_list()

eli5.show_weights(
    permuter,
    top=10,
    feature_names = feature_names
)

# Manually removing any features that a below 0.00001 permutation importance score
cutoff = 0.00001

mask = permuter.feature_importances_ >= cutoff

new_features = X_engineered.columns[mask]

X_trans = X_engineered.copy()
X_trans = X_trans[new_features]
X_test_trans = X_test_engineered.copy()
X_test_trans = X_test_trans[new_features]
print(f'Shape of X_train before removing features: {X_engineered.shape}')
print('Shape after removing features: ', X_trans.shape)

# Think about removing any outliers:

# Remove the most extreme 1% prices,
# the most extreme .1% latitudes, &
# the most extreme .1% longitudes
df = df[(df['price'] >= np.percentile(df['price'], 0.5)) & 
        (df['price'] <= np.percentile(df['price'], 99.5)) & 
        (df['latitude'] >= np.percentile(df['latitude'], 0.05)) & 
        (df['latitude'] < np.percentile(df['latitude'], 99.95)) &
        (df['longitude'] >= np.percentile(df['longitude'], 0.05)) & 
        (df['longitude'] <= np.percentile(df['longitude'], 99.95))]


# Export data
path = '../data/project/'

# Training data
X_enginee.to_csv(path+'X_train_engineered.csv')
y.to_csv(path+'y_train_engineered.csv')

# Test data
X_test_trans.to_csv(path+'X_test_engineered.csv')
