==> notebooks/4_models.ipynb <==
# Imports

import pandas as pd
import pandas_profiling
import numpy as np
import seaborn as sns
import matplotlib.pyplot as plt
from pathlib import Path
from sklearn.model_selection import train_test_split, KFold, cross_val_score, validation_curve, RandomizedSearchCV
from sklearn.ensemble import RandomForestRegressor
from sklearn.pipeline import make_pipeline
from category_encoders import OrdinalEncoder, BinaryEncoder, OneHotEncoder
from sklearn.compose import ColumnTransformer
from sklearn.impute import SimpleImputer
from sklearn.pipeline import make_pipeline, Pipeline
from sklearn.tree import DecisionTreeClassifier
from sklearn.decomposition import PCA
from sklearn.model_selection import GridSearchCV

from sklearn.feature_selection import SelectKBest, f_classif, chi2, SelectPercentile
from xgboost import XGBRegressor
import xgboost as xgb
from sklearn.preprocessing import LabelEncoder
pd.options.display.max_rows = 80


from sklearn.metrics import mean_squared_error as mse

# Importing the data 

data_path = Path('../data/project')
X = pd.read_csv(data_path/'X_train_engineered.csv', index_col = 'Id')
y = pd.read_csv(data_path/'y_train_engineered.csv', index_col = 'Id')
X_test = pd.read_csv(data_path/'X_test_engineered.csv', index_col = 'Id')

my_data = pd.concat([X, y], axis=1)
my_data.head()

# Splitting out data into training and validation sets
from sklearn.model_selection import train_test_split


X_train, X_val, y_train, y_val = train_test_split(X, y, test_size = 0.2)

# Un-log our y_val
y_val = np.exp(y_val)
X_train.shape, X_val.shape, y_train.shape, y_val.shape

# Fitting some models:
* Baseline linear model
* Random Forest
* XGBoost Regression
* Ridge regression

# BASIC LINEAR REGRESSOR WITH NON ENGINEERED DATA

# Baseline average guesser
data_path = Path('../data/project')
df = pd.read_csv(data_path/'train.csv', index_col = 'Id')
df_test = pd.read_csv(data_path/'test.csv', index_col = 'Id')

# Creating our dataset splits
target = 'SalePrice'
X_full = df.drop(columns=target)
y_full = df[target]
X_t, X_v, y_t, y_v = train_test_split(X_full, y_full, test_size = 0.2, random_state = 42)

# Building our predictions
y_pred = pd.Series([y_full.mean()] * len(y_v))

# Getting RMSE
rmse_mean_baseline = mse(y_v, y_pred, squared=False)

print(f'The validation RMSE of my mean baseline model is: {rmse_mean_baseline}')

# LINEAR REGRESSION MODEL WITH NON ENGINEERED DATA

# Baseline linear model

model_lr = Pipeline(steps=[
    ('encoder', OrdinalEncoder()),
    ('imputer', SimpleImputer(strategy='constant', fill_value=0)),
    ('randomforestregressor', LinearRegression(n_jobs = -1))
])
model_lr.fit(X_t, y_t);

# Getting baseline RMSE


y_pred = model_lr.predict(X_v)

# Reversing the log target
#y_pred = np.exp(y_pred)
#y_pred = pd.Series(y_pred)
rmse_baseline = mse(y_v, y_pred, squared=False)

print(f'The validation RMSE of my baseline is: {rmse_baseline}')

# RANDOM FOREST MODEL WITH ENGINEERED DATA

X_train.columns

# Baseline linear model

model_rf = Pipeline(steps=[
    ('randomforestregressor', RandomForestRegressor(n_jobs = -1))
])
model_rf.fit(X_train, y_train);

# Results: 

y_pred = model_rf.predict(X_val)
y_pred_train = model_rf.predict(X_train)

# Reversing the log target
y_pred = np.exp(y_pred)
y_pred_train = np.exp(y_pred_train)

rmse_rf = mse(y_val, y_pred, squared=False)
rmse_rf_train = mse(np.exp(y_train), y_pred_train, squared=False)

print(f'The training RMSE of my Random Forest model is: {rmse_rf_train}')
print(f'The validation RMSE of my Random Forest model is: {rmse_rf}')

# XGBoost regression
from xgboost import XGBRegressor

model_gb = make_pipeline(
    OrdinalEncoder(),
    SimpleImputer(strategy = 'constant', fill_value = 0), 
    XGBRegressor(n_estimators=200, objective='reg:squarederror', n_jobs=-1)
)
model_gb.fit(X_train, y_train)
# Results: 

y_pred = model_gb.predict(X_val)

# Reversing the log target
y_pred = np.exp(y_pred)

rmse_gb = mse(y_val, y_pred, squared=False)

print(f'The validation RMSE of my XGBoost model is: {rmse_gb}')

# Feature selection using feature permutation:


# Hyper paramter tuning:

from scipy.stats import randint, uniform
from sklearn.ensemble import RandomForestRegressor
from sklearn.model_selection import RandomizedSearchCV

param_distributions = { 
    'randomforestregressor__n_estimators': randint(50, 500), 
    'randomforestregressor__max_depth': [5, 10, 15, 20, None], 
    'randomforestregressor__max_features': uniform(0, 1), 
}

search = RandomizedSearchCV(
    model_rf, 
    param_distributions=param_distributions, 
    n_iter=5, 
    cv=2, 
    scoring='neg_mean_absolute_error', 
    verbose=10, 
    return_train_score=True, 
    n_jobs=-1
)

search.fit(X_train, y_train);

print('Best hyperparameters', search.best_params_)
print('Cross-validation MAE', -search.best_score_)
model = search.best_estimator_

y_pred = model.predict(X_val)
y_pred_train = model.predict(X_train)

y_pred = np.exp(y_pred)
y_pred_train = np.exp(y_pred_train)


rmse_rf = mse(y_val, y_pred, squared=False)
rmse_rf_train = mse(np.exp(y_train), y_pred_train, squared=False)

print(f'The training RMSE of my Random Forest model is: {rmse_rf_train}')
print(f'The validation RMSE of my Random Forest model is: {rmse_rf}')

# Creating a predict function that also displays the shapley scores/plot
